Add armi.reactor.grids.thetarz
